refactor(api): extract haversineKm into a shared geo util
The Memories Trips curator (#304) needs the same great-circle distance the
location-inference engine already computes. Issue #304 calls for moving the
helper to a shared util rather than duplicating it: two haversines drifting
apart on an earth-radius constant would make a location suggestion and a trip
boundary disagree about the same coordinates.

location-inference.service.ts re-exports haversineKm, so its importers and its
golden-value spec are untouched.
